Add new 'gerrit project tag create' command and API
This patch adds new 'gerrit project tag create' command with
respective API, that allows to create a tag for a specified
project:

usage: gerrit project tag create [-h] [-f {json,shell,table,value,yaml}]
                                 [-c COLUMN] [--max-width <integer>]
                                 [--fit-width] [--print-empty] [--noindent]
                                 [--prefix PREFIX] -t TAG [-r REVISION]
                                 [-m MESSAGE]
                                 project-identifier

Creates a new tag on the project. If a message is provided in the input, the
tag is created as an annotated tag with the current user as tagger. Signed
tags are not supported.

positional arguments:
  project-identifier    Project identifier.

optional arguments:
  -h, --help            show this help message and exit
  -t TAG, --tag TAG     The name of the tag. The leading refs/tags/ is
                        optional.
  -r REVISION, --revision REVISION
                        The revision to which the tag should point. If not
                        specified, the project's HEAD will be used.
  -m MESSAGE, --message MESSAGE
                        The tag message. When set, the tag will be created as
                        an annotated tag.

e.g.:

   gerrit project tag create some/fake-project --tag v9.0